Reorganized the example module code
