+++ farmdata2_modules/fd2_tabs/fd2_example/info/info.spec.js
